Include saved_report_paging only where needed.
