+++ app/controllers/report_controller.rb
@@ -11,6 +11,7 @@ class ReportController < ApplicationController
   helper ApplicationHelper::ImportExportHelper
   include ReportHelper
   include Mixins::GenericSessionMixin
+  include Mixins::SavedReportPaging
 
   before_action :check_privileges
   before_action :get_session_data
